+++ src/app/components/ui/header/NavBar/index.js
@@ -1,0 +1,3 @@
+import NavBar from './NavBar.jsx'
+
+export default NavBar
