+++ src/app/components/ui/header/NavLogo.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { NavLink } from 'react-router-dom'
-import { MainTitle } from '../../typography'
+import { MainTitle } from '../typography'
 import PropTypes from 'prop-types'
 
 const NavLogo = ({ label }) => {
