+++ src/app/components/ui/temmateCard/index.js
@@ -1,0 +1,3 @@
+import TeammateCard from './TeammateCard'
+
+export default TeammateCard
